random gen, scriptable objs, integration, etc (this should have been like 10 separate commits LOL)

+++ Assets/Scripts/Puzzles/PuzzleData.cs
@@ -5,6 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "PuzzleData", menuName = "ScriptableObjects/PuzzleData")]
 public class PuzzleData : ScriptableObject
 {
-    public float difficulty;
+    // number from 1 to 5
+    public int difficulty;
     public bool solved = false;
 }
